[Task 72232] altered email param usage for curation scripts when being called from cli or rest
