+++ dspace-api/src/main/java/org/dspace/curate/CurationClientOptions.java
@@ -54,7 +54,6 @@ public enum CurationClientOptions {
             "Id (handle) of object to perform task on, or 'all' to perform on whole repository");
         options.addOption("p", "parameter", true, "a task parameter 'NAME=VALUE'");
         options.addOption("q", "queue", true, "name of task queue to process");
-        options.addOption("e", "eperson", true, "email address of curating eperson");
         options.addOption("r", "reporter", true,
             "relative or absolute path to the desired report file. Use '-' to report to console. If absent, no " +
             "reporting");
