+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/scripts/handler/impl/RestDSpaceRunnableHandler.java
@@ -33,7 +33,7 @@ import org.dspace.scripts.factory.ScriptServiceFactory;
 import org.dspace.scripts.handler.DSpaceRunnableHandler;
 import org.dspace.scripts.service.ProcessService;
 import org.dspace.utils.DSpace;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import org.springframework.core.task.TaskExecutor;
 
 /**
  * The {@link DSpaceRunnableHandler} dealing with Scripts started from the REST api
@@ -231,9 +231,8 @@ public class RestDSpaceRunnableHandler implements DSpaceRunnableHandler {
      * @param script    The script to be ran
      */
     public void schedule(DSpaceRunnable script) {
-        ThreadPoolTaskExecutor taskExecutor = new DSpace().getServiceManager()
-                                                          .getServiceByName("dspaceRunnableThreadExecutor",
-                                                                            ThreadPoolTaskExecutor.class);
+        TaskExecutor taskExecutor = new DSpace().getServiceManager()
+                                                .getServiceByName("dspaceRunnableThreadExecutor", TaskExecutor.class);
         Context context = new Context();
         try {
             Process process = processService.find(context, processId);
